Remove package execs
Remove package execs and move RunCmd() to package cmdtmpl.
