+++ internal/cmdtmpl/command_test.go
@@ -2,6 +2,7 @@ package cmdtmpl
 
 import (
 	"context"
+	"path/filepath"
 	"testing"
 	"text/template"
 
@@ -54,6 +55,39 @@ func TestGetCommandList(t *testing.T) {
 		if cl.Name != name {
 			t.Errorf("command list should be %s, got %s", name, cl.Name)
 		}
+	}
+}
+
+// TestRunCmd tests RunCmd.
+func TestRunCmd(t *testing.T) {
+	ctx := context.Background()
+
+	// test not existing
+	dir := t.TempDir()
+	if _, _, err := RunCmd(ctx, filepath.Join(dir, "does/not/exist"), ""); err == nil {
+		t.Errorf("running not existing command should fail: %v", err)
+	}
+
+	// test existing
+	if _, _, err := RunCmd(ctx, "echo", "", "this", "is", "a", "test"); err != nil {
+		t.Errorf("running echo failed: %v", err)
+	}
+
+	// test with stdin
+	if _, _, err := RunCmd(ctx, "echo", "this is a test"); err != nil {
+		t.Errorf("running echo failed: %v", err)
+	}
+
+	// test stdout
+	stdout, stderr, err := RunCmd(ctx, "cat", "this is a test")
+	if err != nil || string(stdout) != "this is a test" {
+		t.Errorf("running echo failed: %s, %s, %v", stdout, stderr, err)
+	}
+
+	// test stderr and error
+	stdout, stderr, err = RunCmd(ctx, "cat", "", "does/not/exist")
+	if err == nil || string(stderr) != "cat: does/not/exist: No such file or directory\n" {
+		t.Errorf("running echo failed: %s, %s, %v", stdout, stderr, err)
 	}
 }
